Fibonacci uses unsigned long long for larger terms

+++ 03-loops/Fibonacci.c
@@ -1,16 +1,16 @@
 #include <stdio.h>
 
 int main() {
-    int n, i;
-    int t1 = 0, t2 = 1, nextTerm;
+    unsigned long long n, i;
+    unsigned long long t1 = 0, t2 = 1, nextTerm;
 
     printf("Enter the number of terms: ");
-    scanf("%d", &n);
+    scanf("%llu", &n);
 
     printf("Fibonacci Series: ");
 
     for (i = 1; i <= n; i++) {
-        printf("%d", t1);
+        printf("%llu", t1);
         if (i < n) {
             printf(", ");
         }
